Lift sheetlet from document to view
