+++ components/experimental/table-document/package.json
@@ -52,7 +52,6 @@
     "@fluidframework/protocol-definitions": "^0.1008.0-0",
     "@fluidframework/runtime-definitions": "^0.21.0",
     "@fluidframework/sequence": "^0.21.0",
-    "@tiny-calc/micro": "0.0.0-alpha.0",
     "debug": "^4.1.1"
   },
   "devDependencies": {
